BackEndHemoLife/requirements.txt
fastapi==0.95.0
uvicorn==0.23.2
sqlalchemy==1.4.46
alembic==1.11.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-decouple==3.8
email-validator==2.0.0